Move ResourceAllocatorPrometheusCallback to callbacks directory
