+++ python/ray/data/context.py
@@ -984,10 +984,10 @@ class DataContext:
         from ray.data._internal.execution.callbacks.insert_issue_detectors import (
             IssueDetectionExecutionCallback,
         )
-        from ray.data._internal.execution.execution_callback import ExecutionCallback
-        from ray.data._internal.execution.resource_allocator_prometheus_callback import (
+        from ray.data._internal.execution.callbacks.resource_allocator_prometheus_callback import (
             ResourceAllocatorPrometheusCallback,
         )
+        from ray.data._internal.execution.execution_callback import ExecutionCallback
 
         classes = [
             ExecutionIdxUpdateCallback,
